recommendation_system/requirements.txt
pandas
numpy
python-dotenv
tqdm
openai
qdrant-client
requests
beautifulsoup4
fastapi
uvicorn
pydantic